% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JK_outlier.R
\name{outlier.jk}
\alias{outlier.jk}
\title{IQR-based Outlier Detection (JMK)}
\usage{
outlier.jk(
  variable,
  print_mode = c("values", "cutoffs"),
  multiplier = 1.5,
  cutoff_type = c("high", "low", "both")
)
}
\arguments{
\item{variable}{A list of numeric values to subject to outlier testing}

\item{print_mode}{Default option "values" will return actual outliers if present; "cutoffs" with  return criteria to determine outliers}

\item{multiplier}{Number multiplied to the IQR to set outlier thresholds; default is 1.5}

\item{cutoff_type}{For use when print mode set to 'cutoffs'. Choose to display the higher ('high') or lower ('low') cutoffs, or "both"}
}
\value{
Returns either a list of values that are outliers (and their relative position in the query), or numbers to use to filter data for outliers
}
\description{
A function to show outliers from a set of numbers as tested using the IQR method, either by displaying the outlier values when present, or returning the upper and lower criteria for deciding if something is an outlier.
}
\examples{
#returns 15 and -10 as outliers
outlier.jk(c(1,2,3,1,3,4,15,-10))

#shows the range values must be in to be deemed 'not an outlier'
outlier.jk(c(1,2,3,1,3,4,15,-10) , print_mode = "cutoffs", cutoff_type = "both")
}
